dev: + dropped lang (the config stays), and uses lepet + drpopped handler helper
