+++ lang/vars.go
@@ -1,7 +1,11 @@
 package lang
 
+import (
+	l "github.com/karincake/ambeng/lepet"
+)
+
 // THe built-in language item list for error messages
-var defaultList = langItem{
+var defaultList = l.LangItem{
 	"required":        "value is required",
 	"integerPositive": "value must be a positive integer",
 	"uuid":            "value must be a valid UUID",
